Minor changes to Game and ParseLine
Minor changes to Game: inventory is now a string vector holding the name of the Item

ParseLine: Improved parsing for interacting with items -- once the player grabs the item, the message "You grabbed the <item>" prints, followed by the item's description.  Also added a constraint not allowing duplicate entries into the inventory.  In addition, the player can type "inventory" at any time and either the list of their inventory prints, or a message stating the have no inventory prints if their inventory is empty.